size_t changes done

+++ exercises/solutions/04_F_advanced/ex15_generic_list/List.hpp
@@ -49,27 +49,27 @@ public:
      */
 	void prependElement(const T &i);			// prepend an element to the beginning of the list
     /**
-     * @name insertElementAt(const T &i, int pos);
+     * @name insertElementAt(const T &i, size_t pos);
      * @brief insert an element at a specific position in the list
      * @author cppp
      * @param i content which will interted
      * @param pos the position
      */
-	void insertElementAt(const T &i, int pos);	// insert an element i at position pos
+	void insertElementAt(const T &i, size_t pos);	// insert an element i at position pos
     /**
      * @name getSize() const;
      * @brief get the current size of the list
      * @author cppp
      */
-	int getSize() const;		// get the number of elements in list
+	size_t getSize() const;		// get the number of elements in list
     /**
-     * @name getNthElement(int n);
+     * @name getNthElement(size_t n);
      * @brief get the content of the n-th list item
      * @author cppp
      * @param n the position of the n-th item
      * @return the content of the n-th listitem as T reference
      */
-	T& getNthElement(int n);	// get content of the n-th element.
+	T& getNthElement(size_t n);	// get content of the n-th element.
     /**
      * @name getFirst();
      * @brief get the content of the first list item
@@ -97,12 +97,12 @@ public:
      */
 	void deleteLast();			// delete last element
     /**
-     * @name deleteAt(int pos);
+     * @name deleteAt(size_t pos);
      * @brief deleting the n-th element out of the list
      * @author cppp
      * @param pos the position at which the listitem will be deleted
      */
-	void deleteAt(int pos);		// delete element at position pos
+	void deleteAt(size_t pos);		// delete element at position pos
     /**
      * @name begin();
      * @brief returns a list iteration pointing to the first listitem
@@ -119,7 +119,7 @@ public:
 	ListIterator<T> end();		// return an iterator pointing to the element after the last one
 private:
 	ListItem<T> *first, *last;	// first and last item pointers (NULL if list is empty)
-	int currentSize;			// current size of the list
+	size_t currentSize;			// current size of the list
 	friend class ListIterator<T>;
 };
